Neutron CI Monitors
* Offline nodes monitor. It goes to Jenkins and iy checks
if there are offline nodes. If so it waits for some time and
check is the nodes back online. Exit code is a number
offline nodes.
* Zuul queue monitor. It goes to Zuul status page, if the
pipeline (specified in parameters) is empty there is a need
to check how long time ago jobs were lunched last time.
If there was no any activities for long time if fails
* Jenkins builds monitor. Checks status of last builds in
jobs history. If there is at least one build with expected status
everything is OK.